requirements.txt
# 热点新闻收集与推送工具依赖项

# 基础库
requests>=2.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
feedparser>=6.0.0

# 异步和并发
asyncio>=3.4.3

# LangChain相关
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.5

# 日志和工具
logging>=0.5.1.2
pathlib>=1.0.1

# 数据处理
json>=2.0.9
pickle-mixin>=1.0.2